Remove OAuthRequestAsync, just OAuthRequest. Request should not know anything about sync vs async
