+++ changelog
@@ -1,3 +1,6 @@
+[SNAPSHOT]
+ * Remove OAuthRequestAsync, just OAuthRequest. Request should know about sync vs async
+
 [3.4.1]
  * Drop deprecated methods
  * Move doktornarabote.ru urls to https (thanks to https://github.com/ezibrov)
