+++ scribejava-apis/src/main/java/com/github/scribejava/apis/service/HHOAuthServiceImpl.java
@@ -1,9 +1,9 @@
 package com.github.scribejava.apis.service;
 
 import com.github.scribejava.core.builder.api.DefaultApi20;
-import com.github.scribejava.core.model.AbstractRequest;
 import com.github.scribejava.core.model.OAuth2AccessToken;
 import com.github.scribejava.core.model.OAuthConfig;
+import com.github.scribejava.core.model.OAuthRequest;
 import com.github.scribejava.core.oauth.OAuth20Service;
 
 public class HHOAuthServiceImpl extends OAuth20Service {
@@ -13,7 +13,7 @@ public class HHOAuthServiceImpl extends OAuth20Service {
     }
 
     @Override
-    public void signRequest(OAuth2AccessToken accessToken, AbstractRequest request) {
+    public void signRequest(OAuth2AccessToken accessToken, OAuthRequest request) {
         request.addHeader("Authorization", "Bearer " + accessToken.getAccessToken());
     }
 }
