+++ scribejava-apis/src/main/java/com/github/scribejava/apis/service/ImgurOAuthServiceImpl.java
@@ -2,10 +2,10 @@ package com.github.scribejava.apis.service;
 
 import com.github.scribejava.apis.ImgurApi;
 import com.github.scribejava.core.builder.api.DefaultApi20;
-import com.github.scribejava.core.model.AbstractRequest;
 import com.github.scribejava.core.model.OAuth2AccessToken;
 import com.github.scribejava.core.model.OAuthConfig;
 import com.github.scribejava.core.model.OAuthConstants;
+import com.github.scribejava.core.model.OAuthRequest;
 import com.github.scribejava.core.oauth.OAuth20Service;
 
 public class ImgurOAuthServiceImpl extends OAuth20Service {
@@ -15,7 +15,9 @@ public class ImgurOAuthServiceImpl extends OAuth20Service {
     }
 
     @Override
-    protected <T extends AbstractRequest> T createAccessTokenRequest(String oauthVerifier, T request) {
+    protected OAuthRequest createAccessTokenRequest(String oauthVerifier) {
+        final DefaultApi20 api = getApi();
+        final OAuthRequest request = new OAuthRequest(api.getAccessTokenVerb(), api.getAccessTokenEndpoint());
         final OAuthConfig config = getConfig();
         request.addBodyParameter(OAuthConstants.CLIENT_ID, config.getApiKey());
         request.addBodyParameter(OAuthConstants.CLIENT_SECRET, config.getApiSecret());
@@ -31,7 +33,7 @@ public class ImgurOAuthServiceImpl extends OAuth20Service {
     }
 
     @Override
-    public void signRequest(OAuth2AccessToken accessToken, AbstractRequest request) {
+    public void signRequest(OAuth2AccessToken accessToken, OAuthRequest request) {
         request.addHeader("Authorization",
                 accessToken == null
                         ? "Client-ID " + getConfig().getApiKey() : "Bearer " + accessToken.getAccessToken());
