+++ scribejava-apis/src/main/java/com/github/scribejava/apis/service/MailruOAuthServiceImpl.java
@@ -7,9 +7,9 @@ import java.util.TreeMap;
 import org.apache.commons.codec.CharEncoding;
 import static org.apache.commons.codec.digest.DigestUtils.md5Hex;
 import com.github.scribejava.core.builder.api.DefaultApi20;
-import com.github.scribejava.core.model.AbstractRequest;
 import com.github.scribejava.core.model.OAuth2AccessToken;
 import com.github.scribejava.core.model.OAuthConfig;
+import com.github.scribejava.core.model.OAuthRequest;
 import com.github.scribejava.core.oauth.OAuth20Service;
 
 public class MailruOAuthServiceImpl extends OAuth20Service {
@@ -19,7 +19,7 @@ public class MailruOAuthServiceImpl extends OAuth20Service {
     }
 
     @Override
-    public void signRequest(OAuth2AccessToken accessToken, AbstractRequest request) {
+    public void signRequest(OAuth2AccessToken accessToken, OAuthRequest request) {
         // sig = md5(params + secret_key)
         request.addQuerystringParameter("session_key", accessToken.getAccessToken());
         request.addQuerystringParameter("app_id", getConfig().getApiKey());
