+++ scribejava-apis/src/main/java/com/github/scribejava/apis/service/TutByOAuthServiceImpl.java
@@ -1,10 +1,10 @@
 package com.github.scribejava.apis.service;
 
 import com.github.scribejava.core.builder.api.DefaultApi20;
-import com.github.scribejava.core.model.AbstractRequest;
 import com.github.scribejava.core.model.OAuth2AccessToken;
 import com.github.scribejava.core.model.OAuthConfig;
 import com.github.scribejava.core.model.OAuthConstants;
+import com.github.scribejava.core.model.OAuthRequest;
 import com.github.scribejava.core.oauth.OAuth20Service;
 
 public class TutByOAuthServiceImpl extends OAuth20Service {
@@ -14,7 +14,7 @@ public class TutByOAuthServiceImpl extends OAuth20Service {
     }
 
     @Override
-    public void signRequest(OAuth2AccessToken accessToken, AbstractRequest request) {
+    public void signRequest(OAuth2AccessToken accessToken, OAuthRequest request) {
         request.addQuerystringParameter(OAuthConstants.TOKEN, accessToken.getAccessToken());
     }
 }
