+++ scribejava-apis/src/test/java/com/github/scribejava/apis/examples/Google20AsyncAHCExample.java
@@ -7,7 +7,7 @@ import com.github.scribejava.httpclient.ahc.AhcHttpClientConfig;
 import com.github.scribejava.core.builder.ServiceBuilder;
 import com.github.scribejava.core.httpclient.HttpClientConfig;
 import com.github.scribejava.core.model.OAuth2AccessToken;
-import com.github.scribejava.core.model.OAuthRequestAsync;
+import com.github.scribejava.core.model.OAuthRequest;
 import com.github.scribejava.core.model.Response;
 import com.github.scribejava.core.model.Verb;
 import com.github.scribejava.core.oauth.OAuth20Service;
@@ -37,88 +37,89 @@ public final class Google20AsyncAHCExample {
                 .setReadTimeout(1_000)
                 .build());
 
-        final OAuth20Service service = new ServiceBuilder()
+        try (OAuth20Service service = new ServiceBuilder()
                 .apiKey(clientId)
                 .apiSecret(clientSecret)
                 .scope("profile") // replace with desired scope
                 .state(secretState)
                 .callback("http://example.com/callback")
                 .httpClientConfig(clientConfig)
-                .build(GoogleApi20.instance());
-        final Scanner in = new Scanner(System.in, "UTF-8");
+                .build(GoogleApi20.instance())) {
+            final Scanner in = new Scanner(System.in, "UTF-8");
 
-        System.out.println("=== " + NETWORK_NAME + "'s OAuth Workflow ===");
-        System.out.println();
-
-        // Obtain the Authorization URL
-        System.out.println("Fetching the Authorization URL...");
-        //pass access_type=offline to get refresh token
-        //https://developers.google.com/identity/protocols/OAuth2WebServer#preparing-to-start-the-oauth-20-flow
-        final Map<String, String> additionalParams = new HashMap<>();
-        additionalParams.put("access_type", "offline");
-        //force to reget refresh token (if usera are asked not the first time)
-        additionalParams.put("prompt", "consent");
-        final String authorizationUrl = service.getAuthorizationUrl(additionalParams);
-        System.out.println("Got the Authorization URL!");
-        System.out.println("Now go and authorize ScribeJava here:");
-        System.out.println(authorizationUrl);
-        System.out.println("And paste the authorization code here");
-        System.out.print(">>");
-        final String code = in.nextLine();
-        System.out.println();
-
-        System.out.println("And paste the state from server here. We have set 'secretState'='" + secretState + "'.");
-        System.out.print(">>");
-        final String value = in.nextLine();
-        if (secretState.equals(value)) {
-            System.out.println("State value does match!");
-        } else {
-            System.out.println("Ooops, state value does not match!");
-            System.out.println("Expected = " + secretState);
-            System.out.println("Got      = " + value);
+            System.out.println("=== " + NETWORK_NAME + "'s OAuth Workflow ===");
             System.out.println();
-        }
 
-        // Trade the Request Token and Verfier for the Access Token
-        System.out.println("Trading the Request Token for an Access Token...");
-        OAuth2AccessToken accessToken = service.getAccessTokenAsync(code, null).get();
-        System.out.println("Got the Access Token!");
-        System.out.println("(if your curious it looks like this: " + accessToken
-                + ", 'rawResponse'='" + accessToken.getRawResponse() + "')");
-
-        System.out.println("Refreshing the Access Token...");
-        accessToken = service.refreshAccessTokenAsync(accessToken.getRefreshToken(), null).get();
-        System.out.println("Refreshed the Access Token!");
-        System.out.println("(if your curious it looks like this: " + accessToken
-                + ", 'rawResponse'='" + accessToken.getRawResponse() + "')");
-        System.out.println();
-
-        // Now let's go and ask for a protected resource!
-        System.out.println("Now we're going to access a protected resource...");
-        while (true) {
-            System.out.println("Paste fieldnames to fetch (leave empty to get profile, 'exit' to stop example)");
+            // Obtain the Authorization URL
+            System.out.println("Fetching the Authorization URL...");
+            //pass access_type=offline to get refresh token
+            //https://developers.google.com/identity/protocols/OAuth2WebServer#preparing-to-start-the-oauth-20-flow
+            final Map<String, String> additionalParams = new HashMap<>();
+            additionalParams.put("access_type", "offline");
+            //force to reget refresh token (if usera are asked not the first time)
+            additionalParams.put("prompt", "consent");
+            final String authorizationUrl = service.getAuthorizationUrl(additionalParams);
+            System.out.println("Got the Authorization URL!");
+            System.out.println("Now go and authorize ScribeJava here:");
+            System.out.println(authorizationUrl);
+            System.out.println("And paste the authorization code here");
             System.out.print(">>");
-            final String query = in.nextLine();
+            final String code = in.nextLine();
             System.out.println();
 
-            final String requestUrl;
-            if ("exit".equals(query)) {
-                break;
-            } else if (query == null || query.isEmpty()) {
-                requestUrl = PROTECTED_RESOURCE_URL;
+            System.out.println("And paste the state from server here. We have set 'secretState'='"
+                    + secretState + "'.");
+            System.out.print(">>");
+            final String value = in.nextLine();
+            if (secretState.equals(value)) {
+                System.out.println("State value does match!");
             } else {
-                requestUrl = PROTECTED_RESOURCE_URL + "?fields=" + query;
+                System.out.println("Ooops, state value does not match!");
+                System.out.println("Expected = " + secretState);
+                System.out.println("Got      = " + value);
+                System.out.println();
             }
 
-            final OAuthRequestAsync request = new OAuthRequestAsync(Verb.GET, requestUrl);
-            service.signRequest(accessToken, request);
-            final Response response = service.execute(request, null).get();
-            System.out.println();
-            System.out.println(response.getCode());
-            System.out.println(response.getBody());
+            // Trade the Request Token and Verfier for the Access Token
+            System.out.println("Trading the Request Token for an Access Token...");
+            OAuth2AccessToken accessToken = service.getAccessTokenAsync(code, null).get();
+            System.out.println("Got the Access Token!");
+            System.out.println("(if your curious it looks like this: " + accessToken
+                    + ", 'rawResponse'='" + accessToken.getRawResponse() + "')");
 
+            System.out.println("Refreshing the Access Token...");
+            accessToken = service.refreshAccessTokenAsync(accessToken.getRefreshToken(), null).get();
+            System.out.println("Refreshed the Access Token!");
+            System.out.println("(if your curious it looks like this: " + accessToken
+                    + ", 'rawResponse'='" + accessToken.getRawResponse() + "')");
             System.out.println();
+
+            // Now let's go and ask for a protected resource!
+            System.out.println("Now we're going to access a protected resource...");
+            while (true) {
+                System.out.println("Paste fieldnames to fetch (leave empty to get profile, 'exit' to stop example)");
+                System.out.print(">>");
+                final String query = in.nextLine();
+                System.out.println();
+
+                final String requestUrl;
+                if ("exit".equals(query)) {
+                    break;
+                } else if (query == null || query.isEmpty()) {
+                    requestUrl = PROTECTED_RESOURCE_URL;
+                } else {
+                    requestUrl = PROTECTED_RESOURCE_URL + "?fields=" + query;
+                }
+
+                final OAuthRequest request = new OAuthRequest(Verb.GET, requestUrl);
+                service.signRequest(accessToken, request);
+                final Response response = service.execute(request, null).get();
+                System.out.println();
+                System.out.println(response.getCode());
+                System.out.println(response.getBody());
+
+                System.out.println();
+            }
         }
-        service.closeAsyncClient();
     }
 }
