+++ scribejava-core/src/main/java/com/github/scribejava/core/model/OAuthRequestAsync.java
@@ -1,15 +1,27 @@
 package com.github.scribejava.core.model;
 
-import java.io.IOException;
+/**
+ *
+ * @deprecated use {@link OAuthRequest}
+ */
+@Deprecated
+public class OAuthRequestAsync extends OAuthRequest {
 
-public class OAuthRequestAsync extends AbstractRequest {
-
+    /**
+     *
+     * @deprecated use {@link OAuthRequest#OAuthRequest(com.github.scribejava.core.model.Verb, java.lang.String) }
+     */
+    @Deprecated
     public OAuthRequestAsync(Verb verb, String url) {
         super(verb, url);
     }
 
-    public interface ResponseConverter<T> {
-
-        T convert(Response response) throws IOException;
+    /**
+     *
+     * @param <T> goal type
+     * @deprecated use {@link OAuthRequest.ResponseConverter}
+     */
+    @Deprecated
+    public interface ResponseConverter<T> extends OAuthRequest.ResponseConverter<T> {
     }
 }
