+++ scribejava-core/src/main/java/com/github/scribejava/core/oauth/OAuth10aService.java
@@ -4,14 +4,12 @@ import java.io.IOException;
 import java.util.Map;
 import java.util.concurrent.Future;
 import com.github.scribejava.core.builder.api.DefaultApi10a;
-import com.github.scribejava.core.model.AbstractRequest;
 import com.github.scribejava.core.model.OAuth1AccessToken;
 import com.github.scribejava.core.model.OAuth1RequestToken;
 import com.github.scribejava.core.model.OAuthAsyncRequestCallback;
 import com.github.scribejava.core.model.OAuthConfig;
 import com.github.scribejava.core.model.OAuthConstants;
 import com.github.scribejava.core.model.OAuthRequest;
-import com.github.scribejava.core.model.OAuthRequestAsync;
 import com.github.scribejava.core.model.Response;
 import com.github.scribejava.core.services.Base64Encoder;
 import com.github.scribejava.core.utils.MapUtils;
@@ -39,9 +37,7 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
     public final OAuth1RequestToken getRequestToken() throws IOException, InterruptedException, ExecutionException {
         final OAuthConfig config = getConfig();
         config.log("obtaining request token from " + api.getRequestTokenEndpoint());
-        final OAuthRequest request = new OAuthRequest(api.getRequestTokenVerb(), api.getRequestTokenEndpoint());
-
-        prepareRequestTokenRequest(request);
+        final OAuthRequest request = prepareRequestTokenRequest();
 
         config.log("sending request...");
         final Response response = execute(request);
@@ -52,14 +48,16 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
         return api.getRequestTokenExtractor().extract(response);
     }
 
+    public final Future<OAuth1RequestToken> getRequestTokenAsync() {
+        return getRequestTokenAsync(null);
+    }
+
     public final Future<OAuth1RequestToken> getRequestTokenAsync(
             OAuthAsyncRequestCallback<OAuth1RequestToken> callback) {
         final OAuthConfig config = getConfig();
         config.log("async obtaining request token from " + api.getRequestTokenEndpoint());
-        final OAuthRequestAsync request
-                = new OAuthRequestAsync(api.getRequestTokenVerb(), api.getRequestTokenEndpoint());
-        prepareRequestTokenRequest(request);
-        return execute(request, callback, new OAuthRequestAsync.ResponseConverter<OAuth1RequestToken>() {
+        final OAuthRequest request = prepareRequestTokenRequest();
+        return execute(request, callback, new OAuthRequest.ResponseConverter<OAuth1RequestToken>() {
             @Override
             public OAuth1RequestToken convert(Response response) throws IOException {
                 return getApi().getRequestTokenExtractor().extract(response);
@@ -67,15 +65,17 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
         });
     }
 
-    protected void prepareRequestTokenRequest(AbstractRequest request) {
+    protected OAuthRequest prepareRequestTokenRequest() {
+        final OAuthRequest request = new OAuthRequest(api.getRequestTokenVerb(), api.getRequestTokenEndpoint());
         final OAuthConfig config = getConfig();
         config.log("setting oauth_callback to " + config.getCallback());
         request.addOAuthParameter(OAuthConstants.CALLBACK, config.getCallback());
         addOAuthParams(request, "");
         appendSignature(request);
+        return request;
     }
 
-    private void addOAuthParams(AbstractRequest request, String tokenSecret) {
+    private void addOAuthParams(OAuthRequest request, String tokenSecret) {
         final OAuthConfig config = getConfig();
         request.addOAuthParameter(OAuthConstants.TIMESTAMP, api.getTimestampService().getTimestampInSeconds());
         request.addOAuthParameter(OAuthConstants.NONCE, api.getTimestampService().getNonce());
@@ -94,10 +94,13 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
     public final OAuth1AccessToken getAccessToken(OAuth1RequestToken requestToken, String oauthVerifier)
             throws IOException, InterruptedException, ExecutionException {
         getConfig().log("obtaining access token from " + api.getAccessTokenEndpoint());
-        final OAuthRequest request = new OAuthRequest(api.getAccessTokenVerb(), api.getAccessTokenEndpoint());
-        prepareAccessTokenRequest(request, requestToken, oauthVerifier);
+        final OAuthRequest request = prepareAccessTokenRequest(requestToken, oauthVerifier);
         final Response response = execute(request);
         return api.getAccessTokenExtractor().extract(response);
+    }
+
+    public final Future<OAuth1AccessToken> getAccessTokenAsync(OAuth1RequestToken requestToken, String oauthVerifier) {
+        return getAccessTokenAsync(requestToken, oauthVerifier, null);
     }
 
     /**
@@ -113,9 +116,8 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
             OAuthAsyncRequestCallback<OAuth1AccessToken> callback) {
         final OAuthConfig config = getConfig();
         config.log("async obtaining access token from " + api.getAccessTokenEndpoint());
-        final OAuthRequestAsync request = new OAuthRequestAsync(api.getAccessTokenVerb(), api.getAccessTokenEndpoint());
-        prepareAccessTokenRequest(request, requestToken, oauthVerifier);
-        return execute(request, callback, new OAuthRequestAsync.ResponseConverter<OAuth1AccessToken>() {
+        final OAuthRequest request = prepareAccessTokenRequest(requestToken, oauthVerifier);
+        return execute(request, callback, new OAuthRequest.ResponseConverter<OAuth1AccessToken>() {
             @Override
             public OAuth1AccessToken convert(Response response) throws IOException {
                 return getApi().getAccessTokenExtractor().extract(response);
@@ -123,18 +125,19 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
         });
     }
 
-    protected void prepareAccessTokenRequest(AbstractRequest request, OAuth1RequestToken requestToken,
-            String oauthVerifier) {
+    protected OAuthRequest prepareAccessTokenRequest(OAuth1RequestToken requestToken, String oauthVerifier) {
+        final OAuthRequest request = new OAuthRequest(api.getAccessTokenVerb(), api.getAccessTokenEndpoint());
         final OAuthConfig config = getConfig();
         request.addOAuthParameter(OAuthConstants.TOKEN, requestToken.getToken());
         request.addOAuthParameter(OAuthConstants.VERIFIER, oauthVerifier);
         config.log("setting token to: " + requestToken + " and verifier to: " + oauthVerifier);
         addOAuthParams(request, requestToken.getTokenSecret());
         appendSignature(request);
+        return request;
     }
 
     @Override
-    public void signRequest(OAuth1AccessToken token, AbstractRequest request) {
+    public void signRequest(OAuth1AccessToken token, OAuthRequest request) {
         final OAuthConfig config = getConfig();
         config.log("signing request: " + request.getCompleteUrl());
 
@@ -162,7 +165,7 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
         return api.getAuthorizationUrl(requestToken);
     }
 
-    private String getSignature(AbstractRequest request, String tokenSecret) {
+    private String getSignature(OAuthRequest request, String tokenSecret) {
         final OAuthConfig config = getConfig();
         config.log("generating signature...");
         config.log("using base64 encoder: " + Base64Encoder.type());
@@ -174,7 +177,7 @@ public class OAuth10aService extends OAuthService<OAuth1AccessToken> {
         return signature;
     }
 
-    private void appendSignature(AbstractRequest request) {
+    private void appendSignature(OAuthRequest request) {
         final OAuthConfig config = getConfig();
         switch (config.getSignatureType()) {
             case Header:
