+++ scribejava-httpclient-ning/src/main/java/com/github/scribejava/httpclient/ning/OAuthAsyncCompletionHandler.java
@@ -1,7 +1,7 @@
 package com.github.scribejava.httpclient.ning;
 
 import com.github.scribejava.core.model.OAuthAsyncRequestCallback;
-import com.github.scribejava.core.model.OAuthRequestAsync;
+import com.github.scribejava.core.model.OAuthRequest;
 import com.github.scribejava.core.model.Response;
 import com.ning.http.client.AsyncCompletionHandler;
 import com.ning.http.client.FluentCaseInsensitiveStringsMap;
@@ -13,10 +13,10 @@ import java.util.Map;
 public class OAuthAsyncCompletionHandler<T> extends AsyncCompletionHandler<T> {
 
     private final OAuthAsyncRequestCallback<T> callback;
-    private final OAuthRequestAsync.ResponseConverter<T> converter;
+    private final OAuthRequest.ResponseConverter<T> converter;
 
     public OAuthAsyncCompletionHandler(OAuthAsyncRequestCallback<T> callback,
-            OAuthRequestAsync.ResponseConverter<T> converter) {
+            OAuthRequest.ResponseConverter<T> converter) {
         this.callback = callback;
         this.converter = converter;
     }
