+++ scribejava-httpclient-okhttp/src/test/java/com/github/scribejava/httpclient/okhttp/OkHttpHttpClientTest.java
@@ -2,7 +2,7 @@ package com.github.scribejava.httpclient.okhttp;
 
 import com.github.scribejava.core.httpclient.HttpClient;
 import com.github.scribejava.core.model.OAuthConfig;
-import com.github.scribejava.core.model.OAuthRequestAsync;
+import com.github.scribejava.core.model.OAuthRequest;
 import com.github.scribejava.core.model.Response;
 import com.github.scribejava.core.model.Verb;
 import com.github.scribejava.core.oauth.OAuth20Service;
@@ -41,7 +41,7 @@ public class OkHttpHttpClientTest {
 
         final HttpUrl baseUrl = server.url("/testUrl");
 
-        final OAuthRequestAsync request = new OAuthRequestAsync(Verb.GET, baseUrl.toString());
+        final OAuthRequest request = new OAuthRequest(Verb.GET, baseUrl.toString());
         final Response response = oAuthService.execute(request, null).get(30, TimeUnit.SECONDS);
 
         assertEquals(expectedResponseBody, response.getBody());
@@ -65,7 +65,7 @@ public class OkHttpHttpClientTest {
         final HttpUrl baseUrl = server.url("/testUrl");
 
         // request with body
-        OAuthRequestAsync request = new OAuthRequestAsync(Verb.POST, baseUrl.toString());
+        OAuthRequest request = new OAuthRequest(Verb.POST, baseUrl.toString());
         request.setPayload(expectedRequestBody);
         Response response = oAuthService.execute(request, null).get(30, TimeUnit.SECONDS);
 
@@ -77,7 +77,7 @@ public class OkHttpHttpClientTest {
 
 
         // request with empty body
-        request = new OAuthRequestAsync(Verb.POST, baseUrl.toString());
+        request = new OAuthRequest(Verb.POST, baseUrl.toString());
         response = oAuthService.execute(request, null).get(30, TimeUnit.SECONDS);
 
         assertEquals(expectedResponseBody, response.getBody());
@@ -99,7 +99,7 @@ public class OkHttpHttpClientTest {
 
         final HttpUrl baseUrl = server.url("/testUrl");
 
-        final OAuthRequestAsync request = new OAuthRequestAsync(Verb.GET, baseUrl.toString());
+        final OAuthRequest request = new OAuthRequest(Verb.GET, baseUrl.toString());
         final Response response = oAuthService.execute(request, null).get(30, TimeUnit.SECONDS);
 
         assertEquals(expectedResponseBody, StreamUtils.getStreamContents(response.getStream()));
